Created a new own file for the sortGiftsCorrectOrder and changed its name. Changed the function as well in index.tsx
